requirements.txt
authlib==1.3.0
Flask==3.1.2
Flask-SocketIO==5.5.1
requests==2.31.0
urllib3==2.1.0
Werkzeug==3.1.3
gunicorn
psycopg2-binary
eventlet==0.40.3
